re-organise and rename methods
